requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-tinymce4-lite==1.7.5
djangorestframework==3.10.2
idna==2.8
jsmin==2.2.2
oauthlib==3.0.2
Pillow==6.1.0
psycopg2==2.8.3
pycodestyle==2.5.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.25.3
